+++ Source/CkPhysics/Public/CkPhysics/Acceleration/CkAcceleration_Fragment_Data.cpp
@@ -1,3 +1,3 @@
-#include "CkVelocity_Fragment_Params.h"
+#include "CkAcceleration_Fragment_Data.h"
 
 // --------------------------------------------------------------------------------------------------------------------
