+++ Source/CkProjectile/Public/CkProjectile/CkProjectile_Fragment_Data.cpp
@@ -1,4 +1,4 @@
-#include "CkProjectile_Fragment_Params.h"
+#include "CkProjectile_Fragment_Data.h"
 
 // --------------------------------------------------------------------------------------------------------------------
