+++ Source/CkProjectile/Public/CkProjectile/CkProjectile_Utils.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "CkProjectile_Fragment_Params.h"
+#include "CkProjectile_Fragment_Data.h"
 
 #include "CkEcs/Handle/CkHandle.h"
