+++ Source/CkRecord/Public/CkRecord/Record/CkRecord_Fragment.h
@@ -3,7 +3,7 @@
 #include "CkEcs/Entity/CkEntity.h"
 #include "CkEcs/Handle/CkHandle.h"
 
-#include "CkRecord/Record/CkRecord_Fragment_Params.h"
+#include "CkRecord/Record/CkRecord_Fragment_Data.h"
 
 // --------------------------------------------------------------------------------------------------------------------
