+++ Source/CkUnreal/Public/CkUnreal/EntityBridge/CkEntityBridge_ConstructionScript.cpp
@@ -1,6 +1,6 @@
 #include "CkEntityBridge_ConstructionScript.h"
 
-#include "CkEntityBridge_Fragment_Params.h"
+#include "CkEntityBridge_Fragment_Data.h"
 
 #include "CkCore/Actor/CkActor_Utils.h"
 #include "CkCore/Algorithms/CkAlgorithms.h"
