+++ Source/CkUnreal/Public/CkUnreal/EntityBridge/CkEntityBridge_Fragment.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "CkEntityBridge_Fragment_Params.h"
+#include "CkEntityBridge_Fragment_Data.h"
 
 #include <variant>
